requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn
streamlit==1.28.1
requests==2.31.0

# Performance optimizations (optional)
uvloop==0.19.0
httptools==0.6.1

# PDF processing
pypdf==3.17.4

# Vector database
chromadb==0.4.18
posthog==3.0.0  # Add this to ensure compatibility

# AI/ML dependencies
# google-generativeai==0.3.2
sentence-transformers
torch>=1.9.0
transformers>=4.20.0

# Environment and configuration
python-dotenv

# Additional utilities
pydantic
python-multipart
psutil